Add a hack to make submodules work
This hack, if enabled, catches exceptions from jgit and uses the git command line utility(has to
be in PATH) to reset the local branch to the tracked branch and update all files. Afterwards
jgit is instructed to re-read the file repository.

Note that enabling the hack can cause loss of data.