Implement the new scm domain event notifications
Supporting the depreciated update() method creates some problems, for example collisions between
manual updates and automatically scheduled ones. The connector also has to make sure that commits
fetched with the automatic update are reported by the depreciated method.

The new startPoller method is called by the tests because some tests do not set up a remote
repository. This should be fixed in the tests.