Do not poll during connector creation
This has several ill effects:
*) Connector creation blocks until the entire git repo is downloaded,
leading to questionable interactivity in opencit.
*) The first ScmUpdateEvent is raised before the caller has a chance to
set up his context. This could still be a problem, maybe we want to
expose the startPoller method via the SCM domain interface.

Furthermore, the race conditions that were the reason for the update()
should not happen because the do_update() method is synchronized.